requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
fuzzywuzzy==0.18.0
idna==3.10
Levenshtein==0.26.0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
RapidFuzz==3.10.0
redis==5.1.1
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3
